Move utility functions for managing a GalaxyInstance to the `common` module
