+++ wft4galaxy/common.py
@@ -2,21 +2,58 @@ import os as _os
 import json as _json
 import uuid as _uuid
 import types as _types
+import logging as _logging
 import operator as _operator
-
 import collections as _collections
 
-# Galaxy ENV variable names
+# bioblend dependencies
 from bioblend import ConnectionError
+from bioblend.galaxy.objects import GalaxyInstance as ObjGalaxyInstance
 
+# Galaxy ENV variable names
 ENV_KEY_GALAXY_URL = "GALAXY_URL"
 ENV_KEY_GALAXY_API_KEY = "GALAXY_API_KEY"
 
-
-import logging as _logging
-
 _log_format = '%(asctime)s %(levelname)s: %(message)s'
 _logger = _logging.getLogger("WorkflowTest")
+
+
+def _get_galaxy_instance(galaxy_url=None, galaxy_api_key=None):
+    """
+    Private utility function to instantiate and configure a :class:`bioblend.GalaxyInstance`
+
+    :type galaxy_url: str
+    :param galaxy_url: the URL of the Galaxy server
+
+    :type galaxy_api_key: str
+    :param galaxy_api_key: a registered Galaxy API KEY
+
+    :rtype: :class:`bioblend.objects.GalaxyInstance`
+    :return: a new :class:`bioblend.objects.GalaxyInstance` instance
+    """
+    # configure `galaxy_url`
+    if galaxy_url is None:
+        if ENV_KEY_GALAXY_URL not in _os.environ:
+            raise TestConfigError("Galaxy URL not defined!  Use --server or the environment variable {} "
+                                  "or specify it in the test configuration".format(ENV_KEY_GALAXY_URL))
+        else:
+            galaxy_url = _os.environ[ENV_KEY_GALAXY_URL]
+
+    # configure `galaxy_api_key`
+    if galaxy_api_key is None:
+        if ENV_KEY_GALAXY_API_KEY not in _os.environ:
+            raise TestConfigError("Galaxy API key not defined!  Use --api-key or the environment variable {} "
+                                  "or specify it in the test configuration".format(ENV_KEY_GALAXY_API_KEY))
+        else:
+            galaxy_api_key = _os.environ[ENV_KEY_GALAXY_API_KEY]
+
+    # initialize the galaxy instance
+    return ObjGalaxyInstance(galaxy_url, galaxy_api_key)
+
+
+def configure_galaxy_instance(galaxy_url, galaxy_api_key):
+    _os.environ[ENV_KEY_GALAXY_URL] = galaxy_url
+    _os.environ[ENV_KEY_GALAXY_API_KEY] = galaxy_api_key
 
 
 def load_comparator(fully_qualified_comparator_function):
@@ -42,6 +79,7 @@ def load_comparator(fully_qualified_comparator_function):
     except:
         _logger.error("Unexpected error: %s", _sys.exc_info()[0])
     return mod
+
 
 class TestConfigError(RuntimeError):
     pass
